+++ include/EnemyShip2.hpp
@@ -6,7 +6,7 @@
 
 class EnemyShip2: public EnemyShip{
 public:
-	EnemyShip2(int arrivalTime, int hp, float atkspd, float xpos, float ypos, float xvel, float yvel,PlayerShip* p);
+	EnemyShip2(int arrivalTime, float hp, float atkspd, float xpos, float ypos, float xvel, float yvel,PlayerShip* p);
 	void draw();
 	void update();
 	void collide();
