+++ include/EnemyShip3.hpp
@@ -6,7 +6,7 @@
 
 class EnemyShip3: public EnemyShip{
 public:
-	EnemyShip3(int arrivalTime, int hp, float atkspd, float xpos, float ypos, float xvel, float yvel,PlayerShip* p);
+	EnemyShip3(int arrivalTime, float hp, float atkspd, float xpos, float ypos, float xvel, float yvel,PlayerShip* p);
 	void draw();
 	void update();
 	void collide();
